requirements.txt
configparser
finnhub-python
pandas
streamlit
yfinance